# backend/public/api-docs/openapi.yml
openapi: 3.0.3
info:
  title: Ma Commune API
  description: |- 
    API pour l'application mobile et le back-office de Ma Commune.
    Cette specification couvre le socle V1 verifie localement :
    signalements, suivi citoyen, notifications, consultations et agenda communal.
  version: 1.6.0
servers:
  - url: /api
    description: Instance locale ou reverse proxy communal
security:
  - bearerAuth: []
paths:
...
  /incidents:
    get:
      summary: Liste des incidents (avec filtres v1.2)
      tags: [Incidents]
      parameters:
        - { name: q, in: query, schema: { type: string }, description: 'Recherche titre/description' }
        - { name: status, in: query, schema: { type: string, enum: [submitted, in_progress, resolved, rejected] } }
        - { name: category_id, in: query, schema: { type: integer } }
        - { name: sort, in: query, schema: { type: string, enum: [created_at, votes_count, updated_at] } }
        - { name: dir, in: query, schema: { type: string, enum: [asc, desc] } }
      responses:
        '200':
          description: Liste des incidents
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Incident' }
    post:
      summary: Créer un incident
      tags: [Incidents]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IncidentCreate' }
      responses:
        '201':
          description: Incident créé
  /incidents/{id}:
    get:
      summary: Détail d'un incident
      tags: [Incidents]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Détail de l'incident
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Incident' }
    put:
      summary: Modifier un incident (auteur, statut=soumis) (v1.2)
      tags: [Incidents]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                address: { type: string }
      responses:
        '200':
          description: Incident modifié

  # ── Votes (v1.1) ──────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Incident:
      type: object
      properties:
        id: { type: integer }
        reference: { type: string }
        title: { type: string }
        description: { type: string }
        address: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        status: { type: string, enum: [submitted, in_progress, resolved, rejected] }
        votes_count: { type: integer }
        user_id: { type: integer }
        category_id: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        photos: { type: array, items: { $ref: '#/components/schemas/Photo' } }
    IncidentCreate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        address: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        category_id: { type: integer }
...
    Photo:
      type: object
      properties:
        id: { type: integer }
        url: { type: string, format: uri }
        file_name: { type: string }
        sort_order: { type: integer }
...